implement view function for authors

+++ src/main/java/com/mprog/springapp/dao/AuthorDao.java
@@ -3,5 +3,5 @@ package com.mprog.springapp.dao;
 import com.mprog.springapp.model.Author;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthorDao extends JpaRepository<Author, Long> {
+public interface AuthorDao extends JpaRepository<Author, Integer> {
 }
